+++ tests/QueryBuilderTest.php
@@ -549,19 +549,19 @@ class QueryBuilderTest extends TestCase
     public function testDates()
     {
         DB::collection('users')->insert([
-            ['name' => 'John Doe', 'birthday' => new UTCDateTime(Date::parse('1980-01-01 00:00:00')->format('Uv'))],
-            ['name' => 'Robert Roe', 'birthday' => new UTCDateTime(Date::parse('1982-01-01 00:00:00')->format('Uv'))],
-            ['name' => 'Mark Moe', 'birthday' => new UTCDateTime(Date::parse('1983-01-01 00:00:00.1')->format('Uv'))],
-            ['name' => 'Frank White', 'birthday' => new UTCDateTime(Date::parse('1960-01-01 12:12:12.1')->format('Uv'))],
+            ['name' => 'John Doe', 'birthday' => new UTCDateTime((int)Date::parse('1980-01-01 00:00:00')->format('Uv'))],
+            ['name' => 'Robert Roe', 'birthday' => new UTCDateTime((int)Date::parse('1982-01-01 00:00:00')->format('Uv'))],
+            ['name' => 'Mark Moe', 'birthday' => new UTCDateTime((int)Date::parse('1983-01-01 00:00:00.1')->format('Uv'))],
+            ['name' => 'Frank White', 'birthday' => new UTCDateTime((int)Date::parse('1960-01-01 12:12:12.1')->format('Uv'))],
         ]);
 
         $user = DB::collection('users')
-            ->where('birthday', new UTCDateTime(Date::parse('1980-01-01 00:00:00')->format('Uv')))
+            ->where('birthday', new UTCDateTime((int)Date::parse('1980-01-01 00:00:00')->format('Uv')))
             ->first();
         $this->assertEquals('John Doe', $user['name']);
 
         $user = DB::collection('users')
-            ->where('birthday', new UTCDateTime(Date::parse('1960-01-01 12:12:12.1')->format('Uv')))
+            ->where('birthday', new UTCDateTime((int)Date::parse('1960-01-01 12:12:12.1')->format('Uv')))
             ->first();
         $this->assertEquals('Frank White', $user['name']);
 
@@ -578,8 +578,8 @@ class QueryBuilderTest extends TestCase
     public function testImmutableDates()
     {
         DB::collection('users')->insert([
-            ['name' => 'John Doe', 'birthday' => new UTCDateTime(Date::parse('1980-01-01 00:00:00')->format('Uv'))],
-            ['name' => 'Robert Roe', 'birthday' => new UTCDateTime(Date::parse('1982-01-01 00:00:00')->format('Uv'))],
+            ['name' => 'John Doe', 'birthday' => new UTCDateTime((int)Date::parse('1980-01-01 00:00:00')->format('Uv'))],
+            ['name' => 'Robert Roe', 'birthday' => new UTCDateTime((int)Date::parse('1982-01-01 00:00:00')->format('Uv'))],
         ]);
 
         $users = DB::collection('users')->where('birthday', '=', new DateTimeImmutable('1980-01-01 00:00:00'))->get();
